// src/app/pages/inscripciones/interfaces/inscripcion.interface.ts
export interface Inscripcion {
    chrClave: string;
    chrNombre: string;
    chrPaterno: string;
    chrMaterno: string;
    chrSexo: string;
    chrTipoSangre: string;
    chrEstadoCivil: string;
    dtFechaNacimiento: string;
    chrLugarNacimiento: string;
    chrDomicilio: string;
    chrColonia: string;
    chrCiudad: string;
    chrEntidadFederativa: string;
    chrCp: string;
    chrTelefonoCelular: string;
    chrTelefonoCasa: string;
    chrCorreoElectronico: string;
    chrTrabajas: string;
    chrEmpresa: string;
    chrPuesto: string;
    chrSector: string;
    chrDireccion: string;
    chrCodigoPostal: string;
    chrColoniaEmpresa: string;
    chrCiudadEmpresa: string;
    chrHorario: string;
    chrTelefonoEmpresa: string;
    chrDependesEconomicamenteDePadres: string;
    chrNombrePadre: string;
    chrTrabajaPadre: string;
    chrEmpresaPadre: string;
    chrPuestoPadre: string;
    chrDireccionEmpresaPadre: string;
    chrEstadoEmpresaPadre: string;
    chrCiudadEmpresaPadre: string;
    chrTelefonoPadre: string;
    chrNombreMadre: string;
    chrTrabajaMadre: string;
    chrEmpresaMadre: string;
    chrPuestoMadre: string;
    chrDireccionEmpresaMadre: string;
    chrEstadoEmpresaMadre: string;
    chrCiudadEmpresaMadre: string;
    chrTelefonoMadre: string;
    chrLicenciatura: string;
    chrClaveTurno: string;
    chrClaveEscuela: string;
    chrPreparatoriaEgreso: string;
    chrDireccionPreparatoria: string;
    chrTerminosYCondiciones: string;
    chrClaveAlumno :string;
    chrCurp: string;
    chrTienesDiscapacidad: string;
    chrEnfermedadCronica: string;
    chrGrupoMinoritario: string;
    chrConQuienVives: string;
    chrModalidadEstudioPreparatoria: string;
    chrPromedioGeneralPreparatoria: string;
    chrParticipacionesReconocimientosPremios: string;
    chrTeniasBecaPrepa: string;
    chrTipoBecaPrepa: string;

}